Desktop Version/requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
dialogflow==0.6.0
google-api-core==1.13.0
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.22.0
idna==2.8
numpy==1.16.4
pandas==0.25.0
protobuf==3.9.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
